dim_web::routes

Module library

Structs

UnmatchedArgs

Functions

library_delete

Method mapped to DELETE /api/v1/library/<id> deletes the library with the supplied id from the path.

library_get_all

Method mapped to GET /api/v1/library returns a list of all libraries in the database

library_get_media

Method mapped to GET /api/v1/library/<id>/media returns all the movies/tv shows that belong to the library with the id supplied. Method can only be accessed by authenticated users.

library_get_one

Method mapped to GET /api/v1/library/<id> returns info about the library with the supplied id. Method can only be accessed by authenticated users.

library_get_unmatched

Method mapped to GET /api/v1/library/<id>/unmatched returns a list of all unmatched medias to be displayed in the library pages.

library_post

Method maps to POST /api/v1/library, it adds a new library to the database, starts a new scanner for it, then dispatches a event to all clients notifying them that a new library has been created. This method can only be accessed by authenticated users. Method returns 200 OK

library_scan

Method mapped to POST /api/v1/library/<id>/scan, it rescans the library with the supplied id.