dim_web::routes::media

Function map_progress

pub async fn map_progress(
    __arg0: State<AppState>,
    __arg1: Path<i64>,
    __arg2: Query<ProgressParams>,
    __arg3: Extension<User>,
) -> Result<impl IntoResponse, Error>

Method mapped to POST /api/v1/media/<id>/progress is used to map progress for a certain media to the user. This is useful for remembering progress for a movie etc.

Arguments

id - id of the media to modify

Query params

offset - offset in seconds