dim_web::routes::media

Function rematch_media_by_id

pub async fn rematch_media_by_id(
    __arg0: State<AppState>,
    __arg1: Path<i64>,
    __arg2: Json<RematchMediaParams>,
) -> Result<impl IntoResponse, Error>

FIXME: Merge this function into rematch_mediafile as theyre functionally the same fucking thing except here we are matching whole media objects rather than mediafiles. This was a different api in the past because the scanner wasnt intelligent enough to decouple and clean up stale media objects but now that it can do that we can just rematch a matched mediafile and it will work as it should.

TODO: Add ability to specify overrides like episode and season ranges.