dim_web::routes::media

Function tmdb_search

pub async fn tmdb_search(
    __arg0: Query<TmdbSearchParams>,
) -> Result<Response, Error>

Method mapped to GET /api/v1/media/tmdb_search is used to quickly search TMDB based on 3 params, one of which is optional. This is used client side in the rematch utility

Arguments

query - the query we want to send to tmdb, ie movie title, tv show title

year - optional parameter specifying the release year of the media we want to look up

media_type - parameter that tells us what media type we are querying, ie movie or tv show