dim_web::routes

Module tv

Functions

delete_episode_by_id

Method mapped to DELETE /api/v1/episode/<id> allows you to delete a episode belonging to some season.

delete_season_by_id

Method mapped to DELETE /api/v1/season/<id> allows you to delete a season for a particular tv show.

get_season_by_id

Method mapped to GET /api/v1/season/<id> returns info about the season by id

get_season_episodes

Method mapped to GET /api/v1/episode/<id> returns information about a episode for a season.

get_tv_seasons

Method mapped to GET /api/v1/tv/<id>/season returns all seasons for TV Show mapped to the id passed in.

patch_episode_by_id

TODO: Move all of these into a unified update interface for media items Method mapped to PATCH /api/v1/episode/<id> lets you patch information about a episode.

patch_season_by_id

Method mapped to PATCH /api/v1/season/<id> allows you to patch in info about the season.