dim_web::routes::websocket

Type Alias WsMessage

pub type WsMessage = Message;

Aliased Type

enum WsMessage {
    Text(String),
    Binary(Vec<u8>),
    Ping(Vec<u8>),
    Pong(Vec<u8>),
    Close(Option<CloseFrame<'static>>),
}

Variants

Text(String)

A text WebSocket message

Binary(Vec<u8>)

A binary WebSocket message

Ping(Vec<u8>)

A ping message with the specified payload

The payload here must have a length less than 125 bytes.

Ping messages will be automatically responded to by the server, so you do not have to worry about dealing with them yourself.

Pong(Vec<u8>)

A pong message with the specified payload

The payload here must have a length less than 125 bytes.

Pong messages will be automatically sent to the client if a ping message is received, so you do not have to worry about constructing them yourself unless you want to implement a unidirectional heartbeat.

Close(Option<CloseFrame<'static>>)

A close message with the optional close frame.