dim_web::tree

Enum Entry

pub enum Entry<T> {
    Directory {
        folder: String,
        files: Vec<Entry<T>>,
    },
    File(T),
}

Represents a entry which can either be a directory with more entries, or a singular file.

Variants

Directory

Fields

folder: String

files: Vec<Entry<T>>

File(T)

Implementations

impl<T> Entry<T>

pub fn new() -> Self

pub fn build_with<U>( values: impl IntoIterator<Item = U>, k: impl Fn(&U) -> Vec<String>, v: impl Fn(&str, U) -> T, ) -> Self

Helper which can turn a collection of values into a tree. The caller must supply a closure which extracts a key from a value.

pub fn insert(&mut self, keys: impl Iterator<Item = impl AsRef<str>>, value: T)

Method inserts a value in the current entry by recursively scanning it based on the keys supplied. The supplied value will be associated with the last key in the collection of keys supplied.

pub fn compress(&mut self)

Changes the root to the first folder that has children.